Get conservativiness features working
For a sentence, its conservativiness is defined to be the probability it came
from a conservative (via an ngram language model trained on "genuine"
conservative statement multiplied by its sentiment).

This feature has the effect of producing a value near 1 if the statement was
positive and "conservativy" and a value near 0 or -1 otherwise.